CIN-79 rate 변경에 따른 수정처리

+++ src/main/java/com/back/cinetalk/user/MyPage/dto/activity/ReviewByUserResponseDTO.java
@@ -17,7 +17,7 @@ public class ReviewByUserResponseDTO {
     private String poster_id;
     private Double star;
     private String content;
-    private int RateCount;
+    private Long RateCount;
     private int RereviewCount;
     private String regDate;
 }
